Update README.md
a. The purpose of this program is to use strings and see how they work. 
b. For Q1, the inputs are the names and numbers. For Q2, the inputs are the original words. 
c. For Q1, the expected outputs are usernames and passwords from information. For Q2, the ouput is to rearrange words to make a new word. 
d. The execution is the template, debugging, and the layout of the strings. 
e. My program could be improved, by making it more efficient, and finding a better way to make the ouputs.